Decreased padding-right on styleToolbar to 15px (#72)
* Decreased padding-right on styleToolbar to 15px

Awesome app! I was checking it out, and noticed that when it is at 322px screen width, the lettering 'Log in' on your nav runs out of room and pushes 'in' below 'Log' (iphone 5 is 320px width). Maybe decreasing the padding-right will fix the issue, or perhaps a media query. Anyways, just thought I would help if i could! Cheers!

* mobile Header fix, modify Grid and margins